Update Ghostscript version to 9.54.0
The previous version (9.27) contains a critical security vulnerability,
which is actively exploited in the wild since almost ~2 years:
https://nvd.nist.gov/vuln/detail/CVE-2019-10216

In the case of this layer, It allows attackers to completely compromise
the Lambda container by uploading EPS files containing malicious
scripts.

Thanks Jens Müller (Twitter: @jensvoid) who found and reported this
vulnerability in a service using this Lamdba Layer.